fix & feat: threshold decimal input, SPC tooltip value, SPC bell notifications, users/archive bug fixes
- LossTracker: threshold inputs now accept decimal minutes (1.5), mm:ss (1:30), or units (1m 28s) with live preview
- deviation_alerts: LimitsIn changed from int to float to support decimal minute thresholds
- QcSpcChart: tooltip now shows actual measured value alongside in-spec/deviated message
- notifications: SPC alerts from QC reports with warnings surfaced in bell icon (last 24h)
- NotificationBell: added spc_alert and qc_report_submitted to WebSocket reload triggers
- archive_service: URL-decode password/username from DATABASE_URL (fixes mysqldump 500 with special chars like @)
- UserManagement: disable role selector when editing own account; omit role from PUT payload to fix 400 error

+++ backend/app/routers/deviation_alerts.py
@@ -39,11 +39,11 @@ class EscalationConfigIn(BaseModel):
 
 
 class LimitsIn(BaseModel):
-    idle: int = 1
-    breakdown: int = 90
-    alarm: int = 30
-    offline: int = 30
-    setting_change: int = 120
+    idle: float = 1.0
+    breakdown: float = 90.0
+    alarm: float = 30.0
+    offline: float = 30.0
+    setting_change: float = 120.0
 
 
 @router.get("/")
